+++ src/test/kotlin/com/group/libraryapp/service/book/BookServiceTest.kt
@@ -166,12 +166,12 @@ class BookServiceTest @Autowired constructor(
 //        val computerDto = results.first{ result -> result.type == BookType.COMPUTER }
 //        val scienceDto = results.first { result -> result.type == BookType.SCIENCE }
 
-        assertCount(results, BookType.COMPUTER, 2)
-        assertCount(results, BookType.SCIENCE, 1)
+        assertCount(results, BookType.COMPUTER, 2L)
+        assertCount(results, BookType.SCIENCE, 1L)
     }
 
     // 개수 세는게 너무 길어질 수 있으니, 아래처럼 메소드로 뺄 수도 있음, 위에처럼하면 깔끔해짐
-    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Int) {
+    private fun assertCount(results: List<BookStatResponse>, type: BookType, count: Long) {
        assertThat(results.first { result -> result.type == type}.count).isEqualTo(count)
     }
